~ Replaced the original energy distribution algorithm
