+++ ui/algorithm.cpp
@@ -9,11 +9,10 @@
 #include "alg/demo/metricsdemo.h"
 #include "alg/demo/tokendemo.h"
 #include "alg/compression.h"
-#include "alg/energydistribution.h"
+#include "alg/energysharing.h"
 #include "alg/infobjcoating.h"
 #include "alg/leaderelection.h"
 #include "alg/shapeformation.h"
-#include "alg/staticenergy.h"
 
 Algorithm::Algorithm(QString name, QString signature)
     : _name(name),
@@ -123,46 +122,35 @@ void CompressionAlg::instantiate(const int numParticles, const double lambda) {
   }
 }
 
-EnergyDistributionAlg::EnergyDistributionAlg()
-    : Algorithm("Energy Distribution", "energydist") {
-  addParameter("# of Particles","19");
-  addParameter("System Type","1");
+EnergySharingAlg::EnergySharingAlg()
+    : Algorithm("Energy Sharing", "energysharing") {
+  addParameter("# Particles", "100");
+  addParameter("System Growth", "0");
   addParameter("Harvest Rate", "0.9");
-  addParameter("Inhibited Rate", "0.0");
-  addParameter("Capacity", "100.0");
-  addParameter("Threshold", "5.0");
-  addParameter("Environment Energy", "5.0");
-  addParameter("GDH", "10.0");
-  addParameter("Signal Speed", "8");
+  addParameter("Capacity", "20.0");
+  addParameter("Threshold", "10.0");
+  addParameter("Source Energy", "1.0");
 }
 
-void EnergyDistributionAlg::instantiate(
-    const int numParticles, const int systemType, const double harvestRate,
-    const double inhibitedRate, const double capacity, const double threshold,
-    const double environmentEnergy,const double GDH,const int signalSpeed) {
-    if(numParticles<1){
-    emit log("number of particles must be positive");
-  } else if (systemType!=0 && systemType!=1){
-    emit log("please choose 0 or 1");
-  } else if (harvestRate <= 0 || harvestRate > 1) {
-    emit log("harvestRate must be in (0,1]", true);
-  } else if (inhibitedRate < 0 || inhibitedRate > harvestRate) {
-    emit log("inhibitedRate must be in [0,harvestRate]", true);
+void EnergySharingAlg::instantiate(int numParticles, const bool isDynamic,
+                                   const double harvestRate,
+                                   const double capacity,
+                                   const double threshold,
+                                   const double sourceEnergy) {
+  if (numParticles <= 0) {
+   emit log("# particles must be > 0", true);
+  } else if (harvestRate <= 0 || harvestRate >= 1) {
+   emit log("harvestRate must be in (0, 1)", true);
   } else if (capacity <= 0) {
-    emit log("capacity must be > 0", true);
-  } else if (threshold <= 0 || threshold >= capacity) {
-    emit log("threshold must be in (0,capacity)", true);
-  } else if (environmentEnergy <= 1 || environmentEnergy > capacity) {
-    emit log("environmentEnergy must be in (1,capacity]", true);
-  } else if (GDH <= 0 || GDH > capacity) {
-    emit log("GDH must be in (0,capacity]", true);
-  } else if (signalSpeed < 1) {
-    emit log("signalSpeed must be >= 1", true);
+   emit log("capacity must be > 0", true);
+  } else if (threshold <= 0 || threshold > capacity) {
+   emit log("threshold must be in (0, capacity]", true);
+  } else if (sourceEnergy <= 0 || sourceEnergy > capacity) {
+   emit log("sourceEnergy must be in (0, capacity]", true);
   } else {
-    emit setSystem(std::make_shared<EnergyDistributionSystem>(
-                     numParticles,systemType,
-                     harvestRate, inhibitedRate, capacity, threshold,
-                     environmentEnergy, GDH, signalSpeed));
+   emit setSystem(std::make_shared<EnergySharingSystem>(
+                    numParticles, isDynamic, harvestRate, capacity, threshold,
+                    sourceEnergy));
   }
 }
 
@@ -229,37 +217,6 @@ void ShapeFormationAlg::instantiate(const int numParticles,
   }
 }
 
-StaticEnergyAlg::StaticEnergyAlg()
-    : Algorithm("Energy Distribution: Static", "staticenergy") {
-  addParameter("# Particles", "100");
-  addParameter("System Growth", "0");
-  addParameter("Harvest Rate", "0.9");
-  addParameter("Capacity", "20.0");
-  addParameter("Threshold", "10.0");
-  addParameter("Source Energy", "1.0");
-}
-
-void StaticEnergyAlg::instantiate(int numParticles, const bool isDynamic,
-                                  const double harvestRate,
-                                  const double capacity, const double threshold,
-                                  const double sourceEnergy) {
-  if (numParticles <= 0){
-    emit log("# particles must be > 0", true);
-  } else if (harvestRate <= 0 || harvestRate >= 1) {
-    emit log("harvestRate must be in (0, 1)", true);
-  } else if (capacity <= 0) {
-    emit log("capacity must be > 0", true);
-  } else if (threshold <= 0 || threshold > capacity) {
-    emit log("threshold must be in (0, capacity]", true);
-  } else if (sourceEnergy <= 0 || sourceEnergy > capacity) {
-    emit log("sourceEnergy must be in (0, capacity]", true);
-  } else {
-    emit setSystem(std::make_shared<StaticEnergySystem>(
-                     numParticles, isDynamic, harvestRate, capacity, threshold,
-                     sourceEnergy));
-  }
-}
-
 AlgorithmList::AlgorithmList() {
   // Demo algorithms.
   _algorithms.push_back(new DiscoDemoAlg());  
@@ -269,11 +226,10 @@ AlgorithmList::AlgorithmList() {
 
   // General algorithms.
   _algorithms.push_back(new CompressionAlg());
-  _algorithms.push_back(new EnergyDistributionAlg());
+  _algorithms.push_back(new EnergySharingAlg());
   _algorithms.push_back(new InfObjCoatingAlg());    
   _algorithms.push_back(new LeaderElectionAlg());
   _algorithms.push_back(new ShapeFormationAlg());
-  _algorithms.push_back(new StaticEnergyAlg());
 }
 
 AlgorithmList::~AlgorithmList() {
